+++ dgisim/tests/test_card/test_weapon_equipment_cards.py
@@ -1,0 +1,37 @@
+"""
+Tests the basic weapon cards which's statuses directly use the template of WeaponEquipmentStatus
+
+Cards covered are: MagicGuide, RavenBow, TravelersHandySword, WhiteIronGreatsword, WhiteTassel
+"""
+import unittest
+
+from .common_imports import *
+
+
+class TestWeaponEquipmentCard(unittest.TestCase):
+    def test_basic_weapons(self):
+        base_game = ACTION_TEMPLATE.factory().f_player1(
+            lambda p1: p1.factory().f_hand_cards(
+                lambda hcs: hcs.add(RavenBow).add(TravelersHandySword)
+            ).build()
+        ).build()
+
+        # tests that Raven Bow cannot be equiped, but THSword can
+        self.assertFalse(RavenBow.loosely_usable(base_game, Pid.P1))
+        self.assertTrue(TravelersHandySword.loosely_usable(base_game, Pid.P1))
+
+        game_state = set_active_player_id(base_game, Pid.P1, 3)
+        game_state = just(game_state.action_step(Pid.P1, CardAction(
+            card=TravelersHandySword,
+            instruction=StaticTargetInstruction(
+                dices=ActualDices({Element.OMNI: 2}),
+                target=StaticTarget(Pid.P1, Zone.CHARACTERS, 3),
+            ),
+        )))
+        game_state = auto_step(game_state)
+
+        game_state = step_skill(game_state, Pid.P1, CharacterSkill.NORMAL_ATTACK)
+        p1ac = game_state.get_player1().just_get_active_character()
+        p2ac = game_state.get_player2().just_get_active_character()
+        self.assertEqual(p2ac.get_hp(), 7)
+        self.assertIn(TravelersHandySwordStatus, p1ac.get_equipment_statuses())